compilable

+++ maplibre_ios/ios/maplibre_ios/Sources/maplibre_ios/FlutterApi.swift
@@ -2,5 +2,5 @@ import Foundation
 import Flutter
 
 @objc public protocol FlutterApi {
-    @objc func createPlatformView(viewId: Int) -> FlutterPlatformView
+    @objc func createPlatformView(viewId: Int64) -> FlutterPlatformView
 }
